app/requirements.txt
asgiref==3.7.2
binary==1.0.0
Django==5.0.1
djangorestframework==3.14.0
numpy==1.26.4
pillow==10.2.0
psycopg==3.1.17
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
gunicorn==22.0.0
